// frontend/src/api/model/documentSectionBase.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * MapStack API
 * AI学習プラットフォーム MapStack のバックエンドAPI
 * OpenAPI spec version: 0.1.0
 */

/**
 * ドキュメントセクション基本クラス
 */
export interface DocumentSectionBase {
  title: string;
  content: string;
}